# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VIRALOS AI - Autonomous Viral Content Operating System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  pending: 'pending',
  scripting: 'scripting',
  voicing: 'voicing',
  finding_clips: 'finding_clips',
  editing: 'editing',
  rendering: 'rendering',
  done: 'done',
  failed: 'failed',
} as const;

export type ProjectPlatform = typeof ProjectPlatform[keyof typeof ProjectPlatform];


export const ProjectPlatform = {
  youtube_shorts: 'youtube_shorts',
  tiktok: 'tiktok',
  reels: 'reels',
  x_clips: 'x_clips',
  all: 'all',
} as const;

export interface Project {
  id: number;
  title: string;
  prompt: string;
  status: ProjectStatus;
  platform: ProjectPlatform;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  duration?: number | null;
  progress?: number;
  createdAt: string;
  updatedAt: string;
}

export type ProjectInputPlatform = typeof ProjectInputPlatform[keyof typeof ProjectInputPlatform];


export const ProjectInputPlatform = {
  youtube_shorts: 'youtube_shorts',
  tiktok: 'tiktok',
  reels: 'reels',
  x_clips: 'x_clips',
  all: 'all',
} as const;

export interface ProjectInput {
  title: string;
  prompt: string;
  platform: ProjectInputPlatform;
}

export interface ProjectUpdate {
  title?: string;
  status?: string;
  progress?: number;
}

export interface Scene {
  index: number;
  description: string;
  emotion?: string;
  duration: number;
}

export interface GeneratedScript {
  id: number;
  projectId: number;
  hook: string;
  script: string;
  emotionCurve?: string[];
  cta: string;
  platformStyle: string;
  scenes?: Scene[];
  createdAt: string;
}

export interface ScriptGenerationInput {
  projectId: number;
  prompt: string;
  platform: string;
  style?: string;
}

export type ClipSource = typeof ClipSource[keyof typeof ClipSource];


export const ClipSource = {
  pexels: 'pexels',
  pixabay: 'pixabay',
  unsplash: 'unsplash',
  runway: 'runway',
  pika: 'pika',
  ai_generated: 'ai_generated',
} as const;

export interface Clip {
  id: number;
  projectId: number;
  source: ClipSource;
  url: string;
  thumbnailUrl: string;
  duration: number;
  query: string;
  /** @nullable */
  emotionTag?: string | null;
  createdAt: string;
}

export interface ClipSearchInput {
  projectId: number;
  query: string;
  emotionTone?: string;
  count?: number;
}

export type AiClipInputProvider = typeof AiClipInputProvider[keyof typeof AiClipInputProvider];


export const AiClipInputProvider = {
  runway: 'runway',
  pika: 'pika',
  stable_video: 'stable_video',
} as const;

export interface AiClipInput {
  projectId: number;
  description: string;
  duration: number;
  provider?: AiClipInputProvider;
}

export type ClipJobStatus = typeof ClipJobStatus[keyof typeof ClipJobStatus];


export const ClipJobStatus = {
  queued: 'queued',
  processing: 'processing',
  done: 'done',
  failed: 'failed',
} as const;

export interface ClipJob {
  id: number;
  projectId: number;
  status: ClipJobStatus;
  provider: string;
  /** @nullable */
  resultUrl?: string | null;
  createdAt: string;
}

export type VoiceoverInputVoiceStyle = typeof VoiceoverInputVoiceStyle[keyof typeof VoiceoverInputVoiceStyle];


export const VoiceoverInputVoiceStyle = {
  motivational_male: 'motivational_male',
  motivational_female: 'motivational_female',
  dramatic_male: 'dramatic_male',
  dramatic_female: 'dramatic_female',
  calm_male: 'calm_male',
  calm_female: 'calm_female',
} as const;

export interface VoiceoverInput {
  projectId: number;
  text: string;
  voiceStyle?: VoiceoverInputVoiceStyle;
  emotionalTone?: string;
}

export type VoiceoverJobStatus = typeof VoiceoverJobStatus[keyof typeof VoiceoverJobStatus];


export const VoiceoverJobStatus = {
  queued: 'queued',
  processing: 'processing',
  done: 'done',
  failed: 'failed',
} as const;

export interface VoiceoverJob {
  id: number;
  projectId: number;
  status: VoiceoverJobStatus;
  voiceStyle: string;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  durationMs?: number | null;
  createdAt: string;
}

export type CaptionInputStyle = typeof CaptionInputStyle[keyof typeof CaptionInputStyle];


export const CaptionInputStyle = {
  dopamine: 'dopamine',
  minimal: 'minimal',
  cinematic: 'cinematic',
  bold: 'bold',
  emoji_heavy: 'emoji_heavy',
} as const;

export interface CaptionInput {
  projectId: number;
  script: string;
  style?: CaptionInputStyle;
}

export interface CaptionWord {
  word: string;
  startMs: number;
  endMs: number;
  highlighted?: boolean;
  /** @nullable */
  emoji?: string | null;
}

export interface CaptionData {
  id: number;
  projectId: number;
  style: string;
  words: CaptionWord[];
  createdAt: string;
}

export type RenderInputTransitionStyle = typeof RenderInputTransitionStyle[keyof typeof RenderInputTransitionStyle];


export const RenderInputTransitionStyle = {
  cinematic: 'cinematic',
  fast_cut: 'fast_cut',
  smooth: 'smooth',
  beat_sync: 'beat_sync',
} as const;

export interface RenderInput {
  projectId: number;
  transitionStyle?: RenderInputTransitionStyle;
  backgroundMusic?: boolean;
  autoZoom?: boolean;
}

export type RenderJobStatus = typeof RenderJobStatus[keyof typeof RenderJobStatus];


export const RenderJobStatus = {
  queued: 'queued',
  processing: 'processing',
  done: 'done',
  failed: 'failed',
} as const;

export interface RenderJob {
  id: number;
  projectId: number;
  status: RenderJobStatus;
  progress?: number;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  outputUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export interface Video {
  id: number;
  projectId: number;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  platform: string;
  durationSec: number;
  /** @nullable */
  fileSizeBytes?: number | null;
  createdAt: string;
}

export interface ExportInput {
  projectId: number;
  platforms: string[];
}

export type ExportJobPlatform = typeof ExportJobPlatform[keyof typeof ExportJobPlatform];


export const ExportJobPlatform = {
  youtube_shorts: 'youtube_shorts',
  tiktok: 'tiktok',
  reels: 'reels',
  x_clips: 'x_clips',
} as const;

export type ExportJobStatus = typeof ExportJobStatus[keyof typeof ExportJobStatus];


export const ExportJobStatus = {
  queued: 'queued',
  processing: 'processing',
  done: 'done',
  failed: 'failed',
} as const;

export interface ExportJob {
  id: number;
  projectId: number;
  platform: ExportJobPlatform;
  status: ExportJobStatus;
  /** @nullable */
  downloadUrl?: string | null;
  width?: number;
  height?: number;
  /** @nullable */
  fileSizeBytes?: number | null;
  createdAt: string;
}

export interface DashboardStats {
  totalProjects: number;
  videosGenerated: number;
  exportsReady: number;
  inProgress: number;
  successRate: number;
  totalDurationSec?: number;
}

export interface PlatformStat {
  platform: string;
  count: number;
  successRate: number;
}

export interface Trend {
  id: number;
  topic: string;
  platform: string;
  score: number;
  category: string;
  /** @nullable */
  emoji?: string | null;
  createdAt: string;
}

export type ListProjectsParams = {
status?: string;
platform?: string;
};

export type ListTrendsParams = {
platform?: string;
category?: string;
};