# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VIRALOS AI - Autonomous Viral Content Operating System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Script Generation ─────────────────────────────────────────────────────
  /scripts/generate:
    post:
      operationId: generateScript
      tags: [scripts]
      summary: Generate a viral script from a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptGenerationInput"
      responses:
        "200":
          description: Generated script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedScript"
  /scripts/{projectId}:
    get:
      operationId: getScript
      tags: [scripts]
      summary: Get script for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedScript"

  # ── Clips ────────────────────────────────────────────────────────────────
  /clips/search:
    post:
      operationId: searchClips
      tags: [clips]
      summary: Search for clips matching a query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClipSearchInput"
      responses:
        "200":
          description: Clips found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clip"
  /clips/generate:
    post:
      operationId: generateAiClip
      tags: [clips]
      summary: Generate an AI video clip as fallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiClipInput"
      responses:
        "200":
          description: Generated clip job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClipJob"
  /clips/{projectId}:
    get:
      operationId: listProjectClips
      tags: [clips]
      summary: List clips assigned to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Clips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clip"

  # ── Voiceover ────────────────────────────────────────────────────────────
...
  /captions/generate:
    post:
      operationId: generateCaptions
      tags: [captions]
      summary: Generate animated captions for a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptionInput"
      responses:
        "200":
          description: Caption data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptionData"
  /captions/{projectId}:
    get:
      operationId: getCaptions
      tags: [captions]
      summary: Get captions for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Caption data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptionData"

  # ── Video Render / Edit ───────────────────────────────────────────────────
...
  /exports/generate:
    post:
      operationId: generateExports
      tags: [exports]
      summary: Generate platform-ready exports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: Export jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExportJob"
  /exports/{projectId}:
    get:
      operationId: listExports
      tags: [exports]
      summary: List exports for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExportJob"

  # ── Analytics ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, title, prompt, status, platform, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        prompt:
          type: string
        status:
          type: string
          enum: [pending, scripting, voicing, finding_clips, editing, rendering, done, failed]
        platform:
          type: string
          enum: [youtube_shorts, tiktok, reels, x_clips, all]
        thumbnailUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        progress:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [title, prompt, platform]
      properties:
        title:
          type: string
        prompt:
          type: string
        platform:
          type: string
          enum: [youtube_shorts, tiktok, reels, x_clips, all]
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        progress:
          type: integer
    GeneratedScript:
      type: object
      required: [id, projectId, hook, script, cta, platformStyle, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        hook:
          type: string
        script:
          type: string
        emotionCurve:
          type: array
          items:
            type: string
        cta:
          type: string
        platformStyle:
          type: string
        scenes:
          type: array
          items:
            $ref: "#/components/schemas/Scene"
        createdAt:
          type: string
          format: date-time
    Scene:
      type: object
      required: [index, description, duration]
      properties:
        index:
          type: integer
        description:
          type: string
        emotion:
          type: string
        duration:
          type: integer
    ScriptGenerationInput:
      type: object
      required: [projectId, prompt, platform]
      properties:
        projectId:
          type: integer
        prompt:
          type: string
        platform:
          type: string
        style:
          type: string
    Clip:
      type: object
      required: [id, projectId, source, url, thumbnailUrl, duration, query, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        source:
          type: string
          enum: [pexels, pixabay, unsplash, runway, pika, ai_generated]
        url:
          type: string
        thumbnailUrl:
          type: string
        duration:
          type: integer
        query:
          type: string
        emotionTag:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ClipSearchInput:
      type: object
      required: [projectId, query]
      properties:
        projectId:
          type: integer
        query:
          type: string
        emotionTone:
          type: string
        count:
          type: integer
    AiClipInput:
      type: object
      required: [projectId, description, duration]
      properties:
        projectId:
          type: integer
        description:
          type: string
        duration:
          type: integer
        provider:
          type: string
          enum: [runway, pika, stable_video]
    ClipJob:
      type: object
      required: [id, projectId, status, provider, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        status:
          type: string
          enum: [queued, processing, done, failed]
        provider:
          type: string
        resultUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    CaptionInput:
      type: object
      required: [projectId, script]
      properties:
        projectId:
          type: integer
        script:
          type: string
        style:
          type: string
          enum: [dopamine, minimal, cinematic, bold, emoji_heavy]
    CaptionData:
      type: object
      required: [id, projectId, style, words, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        style:
          type: string
        words:
          type: array
          items:
            $ref: "#/components/schemas/CaptionWord"
        createdAt:
          type: string
          format: date-time
    CaptionWord:
      type: object
      required: [word, startMs, endMs]
      properties:
        word:
          type: string
        startMs:
          type: integer
        endMs:
          type: integer
        highlighted:
          type: boolean
        emoji:
          type: ["string", "null"]
...
    ExportInput:
      type: object
      required: [projectId, platforms]
      properties:
        projectId:
          type: integer
        platforms:
          type: array
          items:
            type: string
    ExportJob:
      type: object
      required: [id, projectId, platform, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        platform:
          type: string
          enum: [youtube_shorts, tiktok, reels, x_clips]
        status:
          type: string
          enum: [queued, processing, done, failed]
        downloadUrl:
          type: ["string", "null"]
        width:
          type: integer
        height:
          type: integer
        fileSizeBytes:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...