# lib/api-zod/src/generated/types/clipSearchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VIRALOS AI - Autonomous Viral Content Operating System
 * OpenAPI spec version: 0.1.0
 */

export interface ClipSearchInput {
  projectId: number;
  query: string;
  emotionTone?: string;
  count?: number;
}